Large refactoring - WIP
Untested/unsafe/unstable/unreliable commit
