+++ src/settings.ts
@@ -1,0 +1,4 @@
+
+
+
+export const rootdir = __dirname;
