+++ src/utils/fsutils.ts
@@ -2,14 +2,7 @@
 /* eslint-disable global-require */
 /* eslint-disable import/no-dynamic-require */
 import fs from "fs";
-import Discord, {
-  ApplicationCommandPermissionData,
-  Client,
-  CommandInteraction,
-  MessageEmbed,
-  Snowflake,
-  TextChannel,
-} from "discord.js";
+import Discord from "discord.js";
 import path from "path";
 import { BotCommand, BotConfig, BotModule, Offerings, SupportedDepartment } from "./interface";
 
@@ -63,57 +56,12 @@ const botfiles = {
       .filter((filename) => filename.endsWith(".js")),
 };
 
-const fetchConfigChannels = async (
-  client: Client,
-  ...serverTuples: [{ [key: string]: Snowflake }, Map<string, TextChannel>][]
-): Promise<void> => {
-  serverTuples.forEach(([cfgChannels, channelMap]) => {
-    Object.entries(cfgChannels).forEach(async ([channelName, channelID]) => {
-      channelMap.set(channelName, (await client.channels.fetch(channelID)) as TextChannel);
-    });
-  });
-};
-
 const config = {
   load: (): BotConfig => JSON.parse(fs.readFileSync(configPath, "utf8")),
   save: async (cfg: BotConfig): Promise<void> => {
     await fs.promises.writeFile(configPath, JSON.stringify(cfg, null, 2));
     console.log("Saved config.json");
   },
-};
-
-const canExecuteModule = (
-  cfg: BotConfig,
-  module: BotModule | undefined,
-  eventGuildId: Snowflake | undefined
-): boolean => !!eventGuildId && !!module?.state && module.guilds.some((srv) => cfg.servers[srv].id === eventGuildId);
-
-const getRuleEmbedBase = (interaction: CommandInteraction): MessageEmbed =>
-  new MessageEmbed()
-    .setTitle(`${interaction.guild?.name} Kurallar`)
-    .setTimestamp()
-    .setFooter(`Teşekkürler.`)
-    .setColor("ORANGE")
-    .setThumbnail(interaction.guild?.iconURL() || "");
-
-const permissions = {
-  getOwner: (cfg: BotConfig): ApplicationCommandPermissionData => ({
-    id: cfg.ownerId,
-    type: "USER",
-    permission: true,
-  }),
-  getAdmins: (cfg: BotConfig): ApplicationCommandPermissionData[] => [
-    {
-      id: cfg.servers.cs.roles.admin,
-      type: "ROLE",
-      permission: true,
-    },
-    {
-      id: cfg.servers.dh.roles.admin,
-      type: "ROLE",
-      permission: true,
-    },
-  ],
 };
 
 const loadOfferings = (): Offerings => {
@@ -132,10 +80,6 @@ const loadOfferings = (): Offerings => {
 
 export default {
   botfiles,
-  fetchConfigChannels,
   config,
-  canExecuteModule,
-  getRuleEmbedBase,
-  permissions,
   loadOfferings,
 };
